refactor: exlcude dropped column/non support type from snapshot column stats
